Create template for moving_average and refactor simple_coming_average class
